parser-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-docx==1.2.0
PyMuPDF==1.26.4
pdfplumber==0.11.7
